Feature/issue 375 (#376)
* Groups: move _AddImageWidget in GroupWidgets.dart. Rename it to "GroupAddImageWidget" so that it could be reused [#375]

* Possibility for adding group image when creating group [#375]

* Update CHANGELOG.md
